+++ abacusnbody/hod/AbacusHOD/GRAND_HOD.py
@@ -13,6 +13,7 @@ import os
 import sys
 import time
 from pathlib import Path
+import pkgutil
 import math
 from math import erfc
