+++ abacusnbody/hod/chains/test/test.txt
@@ -1,0 +1,12 @@
+13.318661133034272	14.332600043624685	0.32735883090217116
+13.313194918586513	14.284158835268995	0.35474323459978696
+13.367383872771464	14.319404181224613	0.29681531058020266
+13.20817511101602	14.136182880825242	0.3593040844847063
+13.178947960668657	14.290856153612472	0.20438571803064126
+13.338421926558548	14.280457479530362	0.24430677165319292
+13.295262997465686	14.333234320072606	0.2928515492156885
+13.28083728773051	14.325079302522962	0.3395560322309805
+13.319998396980104	14.242926367975176	0.20347502550690735
+13.173293919347904	14.246307585307664	0.3337818747733378
+13.301704601727232	14.416718801615568	0.29976037363239566
+13.215865320556574	14.291566687505933	0.3404635698009301
