+++ abacusnbody/hod/chains/test/testprob.txt
@@ -1,0 +1,12 @@
+-30041733.10447291
+-26871203.384336863
+-61254635.46651569
+-23401066.352093223
+-35744805.54020346
+-93841569.42605467
+-35276262.627802044
+-21054879.348440554
+-130566173.09430802
+-14902725.006722124
+-24431122.29526871
+-15540124.831286471
