+++ abacusnbody/hod/run_hod.py
@@ -15,7 +15,7 @@ import yaml
 import numpy as np
 import argparse
 
-from .abacus_hod import AbacusHOD
+from AbacusHOD.abacus_hod import AbacusHOD
 
 DEFAULTS = {}
 DEFAULTS['path2config'] = 'config/abacus_hod.yaml'
@@ -41,7 +41,7 @@ def main(path2config):
     
     # throw away run for jit to compile, write to disk
     mock_dict = newBall.run_hod(newBall.tracers, want_rsd, write_to_disk)
-    # xirppi = newBall.compute_xirppi(mock_dict, rpbins, pimax, pi_bin_size)
+    xirppi = newBall.compute_xirppi(mock_dict, rpbins, pimax, pi_bin_size)
 
     # run the fit 10 times for timing
     for i in range(10):
